requirements.txt
imap-tools
pandas
openai
python-dotenv
requests
snowflake-connector-python
pyarrow
